Issue #37 Add Get by id for Terms, Weekdays, Slots

See merge request stefan-j-98/ssa2020-back-hypnotized-chicken!37

+++ SSA2020-Back-Hypnotized-Chicken/SSA2020-Back-Hypnotized-Chicken.API/Controllers/DepartmentsController.cs
@@ -31,7 +31,7 @@ namespace SSA2020_Back_Hypnotized_Chicken.API.Controllers
             return Ok(departmentMapResult);
         }
         [HttpGet("{id}")]
-        public DepartmentDTO GetById(int id)
+        public DepartmentDTO GetById(short id)
         {
             var department = UnitOfWork.DepartmentsRepository.GetDepartmentById(id);
